Basic setup for static builds for GitHub Pages
I am very sure that there are still some things missing for this to
work, but at least I now have static builds that can be served without a
problem.

I am planning on using a custom domain for GitHub Pages, so configuring
stuff for the subdirectoy is not a concern.