Removed the height restriction for printing
I used this to debug the content height so that it fits onto a single A4
page.

I have the content fitted. So I don't need it to be in code like this
any longer. I can add it back on demand. But I want to keep it out of
the code.